qcacld-2.0: Send peer delete wmi cmd before peer detach
In wma_remove_peer, sometimes when peer detach is in progress,
and it has set the peer as invalid, but WMI_PEER_DELETE_CMDID
is still not sent to FW, so msdu for the peer still can be
received. In htt_rx_amsdu_rx_in_order_pop_ll, it will print
a lot of error log, which will cause watch dog.

So the change is sending peer delete cmd to FW firstly, then
processing peer detach in host.

Change-Id: Ia26aa5e673a48ca1d9ccf4f184d0bfab0d6e98ef
CRs-Fixed: 1070216